post/ comment service 인터페이스를 위한 디렉토리 생성 : refactoring
